Updated known issues with Internet Explorer 11
Most people are unable to install Internet Explorer 11 via Windows Update or the offline Installer. This is because the .cab files that contain the needed updates for Internet Explorer 11 can't be extracted properly.  The file was updated with the steps to determine if this is the issue, and how to fix this issue. Also dism shows most of the times an error when adding the packages. Instructions where added to fix this. If needed, expand as required the steps, to make them more dummy proof. This is where I found the fix for this issue -> https://jingyangli.wordpress.com/2015/08/18/ie11-install-problem-neutral-package-installation-failed-exit-code-0x00003715-14101-and-setup-exit-code-0x00009c59-40025/